+++ Source/Sitecore.FakeDb/Pipelines/ReleaseFakeDb/ReleasePipelineWatcher.cs
@@ -1,10 +1,13 @@
 ﻿namespace Sitecore.FakeDb.Pipelines.ReleaseFakeDb
 {
+  using Sitecore.Pipelines;
+
   public class ReleasePipelineWatcher
   {
     public virtual void Process(DbArgs args)
     {
       args.Db.PipelineWatcher.Dispose();
+      CorePipelineFactory.ClearCache();
     }
   }
 }