+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ﻿0.16.1
 * [FIX] Fixed issue when template data might be lost
+* [FIX] Clean CorePipelineFactory cache on db disposing
 
 0.16.0
 * [NEW] Implemented security settings inheritance. Ingroduced "__Secutiry" standard field